requirements.txt
# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# AI/ML & RAG
openai>=1.12.0
langchain>=0.1.0
langchain-core>=0.1.0
langchain-ollama>=0.1.0
langgraph>=0.0.20

# Vector Database
# Note: Use 'pinecone' not 'pinecone-client' (renamed package)
pinecone>=3.0.0

# Database
sqlalchemy>=2.0.25
alembic>=1.13.1

# Redis for Session Management
redis>=5.0.1

# Document Processing
pypdf>=4.0.0

# Configuration
pydantic>=2.5.3
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
email-validator>=2.0.0

# Utilities
httpx>=0.26.0

